Refactor main.go to move CLI logic to maprobe package
Move command parsing and execution logic from cmd/maprobe/main.go to maprobe.Main() function. The main.go is now simplified to handle context, signals, logging setup, and AWS Lambda detection only. This improves code organization and testability.
